make test simplier
